remove Users_2.java and add if exist in User class edit InventoryEmployee.java edit Marketer.java edit Sales.java edit DamagedItem.java edit SalesReturn.java edit GlobalConnection.java
